+++ src/components/Field/Field.js
@@ -18,12 +18,7 @@ import Plot from '../Plot'
 import QuickSelect from '../QuickSelect'
 import { fieldMode } from '../../enums'
 import tools from '../../data/tools'
-import {
-  doesInventorySpaceRemain,
-  farmProductsSold,
-  levelAchieved,
-  nullArray,
-} from '../../utils'
+import { doesInventorySpaceRemain, levelAchieved, nullArray } from '../../utils'
 import { getLevelEntitlements } from '../../utils/getLevelEntitlements'
 
 import './Field.sass'
